Split main file into modules
Fixed a crash when multitwitch was passed a user without a twitch
