+++ DirectoryBot.js
@@ -1,8 +1,9 @@
 const Discord = require('discord.js');
-const { DateTime } = require("luxon");
 const fs = require('fs');
-var chrono = require('chrono-node');
 var encrypter = require('crypto-js');
+var timeModule = require('./DirectoryBot_TimeModule.js');
+var twitchModule = require('./DirectoryBot_TwitchModule.js');
+
 const client = new Discord.Client();
 
 class FriendCode {
@@ -108,11 +109,11 @@ client.on('message', (receivedMessage) => {
                     if (helpOverloads.includes(arguments["words"][0])) {
                         helpCommand(arguments, receivedMessage);
                     } else if (convertOverloads.includes(arguments["words"][0])) {
-                        convertCommand(arguments, receivedMessage);
+                        timeModule.convertCommand(arguments, receivedMessage);
                     } else if (countdownOverloads.includes(arguments["words"][0])) {
-                        countdownCommand(arguments, receivedMessage);
+                        timeModule.countdownCommand(arguments, receivedMessage);
                     } else if (multistreamOverloads.includes(arguments["words"][0])) {
-                        multistreamCommand(arguments, receivedMessage);
+                        twitchModule.multistreamCommand(arguments, receivedMessage, userDictionary);
                     } else if (recordOverloads.includes(arguments["words"][0])) {
                         recordCommand(arguments, receivedMessage);
                     } else if (lookupOverloads.includes(arguments["words"][0])) {
@@ -234,97 +235,6 @@ Syntax: \`@DirectoryBot setplatformrole (platform) (role)\``)
             receivedMessage.author.send(`You need a role with administrator privileges or the role ${receivedMessage.guild.roles.get(opRole)} to view operator commands.`);
         }
     }
-}
-
-
-function convertCommand(arguments, receivedMessage) {
-    var timeText = "";
-    var startTimezone = "";
-    var resultTimezone;
-
-    //TODO check against convertOverloads here to adjust argument indexing after shortcut has been implemented
-
-    if (arguments["userMentions"].length == 1) {
-        for (var i = 0; i < arguments["words"].length; i++) {
-            if (arguments["words"][i] == "in") {
-                startTimezone = arguments["words"][i + 1]
-                i++;
-            } else if (arguments["words"][i] == "for") {
-                break;
-            } else if (arguments["words"][i] != "convert") {
-                timeText += arguments["words"][i] + " ";
-            }
-        }
-        resultTimezone = userDictionary[arguments["userMentions"][0].id]["timezone"];
-    } else {
-        for (var i = 0; i < arguments["words"].length; i++) {
-            if (arguments["words"][i] == "in") {
-                startTimezone = arguments["words"][i + 1]
-                i++;
-            } else if (arguments["words"][i] == "to") {
-                resultTimezone = arguments["words"][i + 1];
-                break;
-            } else if (arguments["words"][i] != "convert") {
-                timeText += arguments["words"][i] + " ";
-            }
-        }
-    }
-    if (startTimezone == "") {
-        if (userDictionary[receivedMessage.author.id] && userDictionary[receivedMessage.author.id]["timezone"].value) {
-            startTimezone = userDictionary[receivedMessage.author.id]["timezone"].value;
-        } else {
-            receivedMessage.author.send(`Please either specifiy a timezone or record your default with \`@DirectoryBot record timezone (timezone)\`.`);
-            return;
-        }
-    }
-    timeText += "(" + startTimezone + ")";
-
-    var inputTime = new chrono.parse(timeText); //BUG chrono misparses "9:00 PDT" as "9:00 PM"
-    var dateTimeObject = DateTime.fromJSDate(inputTime[0].start.date());
-    dateTimeObject.setZone(resultTimezone);
-
-    if (arguments["userMentions"].length == 1) {
-        receivedMessage.channel.send(`${arguments["words"][1]} in ${startTimezone} is ${dateTimeObject.toLocaleString(DateTime.TIME_SIMPLE)} for ${arguments["userMentions"][0]}.`);
-    } else {
-        receivedMessage.channel.send(`${arguments["words"][1]} in ${startTimezone} is ${dateTimeObject.toLocaleString(DateTime.TIME_SIMPLE)} in ${resultTimezone}.`);
-    }
-}
-
-
-function countdownCommand(arguments, receivedMessage) {
-    var timeText = "";
-    for (var i = 0; i < arguments["words"].length; i++) {
-        if (arguments["words"][i] != "countdown") {
-            timeText += arguments["words"][i] + " ";
-        }
-    }
-
-    var inputTime = new chrono.parse(timeText);
-    var dateTimeObject = DateTime.fromJSDate(inputTime[0].start.date());
-    var countdown = dateTimeObject.diffNow("minutes").toString();
-    countdown = countdown.replace(/[a-zA-Z]/g, '');
-    countdown = parseInt(countdown);
-
-    receivedMessage.channel.send(`${arguments["words"][1]} is about ${countdown} minutes from now.`);
-}
-
-
-function multistreamCommand(arguments, receivedMessage) {
-    var url = "https://multistre.am/";
-    var layout = arguments["words"][1];
-
-    for (var i = 0; i < arguments["userMentions"].length; i++) {
-        if (!userDictionary[arguments["userMentions"][i].id]["twitch"].value) {
-            receivedMessage.channel.send(`${arguments[i]} does not have a Twitch account logged with **DirectoryBot**.`)
-            return;
-        }
-        url += userDictionary[arguments["userMentions"][i].id]["twitch"].value + "/";
-    }
-    if (layout) {
-        url += "layout" + layout;
-    }
-
-    receivedMessage.channel.send(`Here's the multistream link: ${url}`);
 }
